// src/lib/api/models/PlanInfo.ts
/* tslint:disable */
/* eslint-disable */
/**
 * LUCO MINTOS
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface PlanInfo
 */
export interface PlanInfo {
    /**
     * 
     * @type {string}
     * @memberof PlanInfo
     */
    planName: string;
    /**
     * 
     * @type {string}
     * @memberof PlanInfo
     */
    smsCost: string;
    /**
     * 
     * @type {string}
     * @memberof PlanInfo
     */
    description: string;
    /**
     * 
     * @type {Array<string>}
     * @memberof PlanInfo
     */
    features: Array<string>;
}

/**
 * Check if a given object implements the PlanInfo interface.
 */
export function instanceOfPlanInfo(value: object): value is PlanInfo {
    if (!('planName' in value) || value['planName'] === undefined) return false;
    if (!('smsCost' in value) || value['smsCost'] === undefined) return false;
    if (!('description' in value) || value['description'] === undefined) return false;
    if (!('features' in value) || value['features'] === undefined) return false;
    return true;
}

export function PlanInfoFromJSON(json: any): PlanInfo {
    return PlanInfoFromJSONTyped(json, false);
}

export function PlanInfoFromJSONTyped(json: any, ignoreDiscriminator: boolean): PlanInfo {
    if (json == null) {
        return json;
    }
    return {
        
        'planName': json['plan_name'],
        'smsCost': json['sms_cost'],
        'description': json['description'],
        'features': json['features'],
    };
}

export function PlanInfoToJSON(json: any): PlanInfo {
    return PlanInfoToJSONTyped(json, false);
}

export function PlanInfoToJSONTyped(value?: PlanInfo | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'plan_name': value['planName'],
        'sms_cost': value['smsCost'],
        'description': value['description'],
        'features': value['features'],
    };
}